Only sort histogram buckets if unsorted
Summary: go's `sort.Sort` function will occasionally modify the data
that is passed into it even if that data is already sorted.
(https://play.golang.org/p/QJ8kiJzzVG)

For tally, this means that if a bucket slice is passed into tally, and
tally calls `sort.Sort` on that slice, the slice will get modified inline
(even if the list is already sorted).

In practice, this means that if a user is using a single slice of buckets
to create multiple tally histograms, there is a race condition where the
shared bucket object could be modified mid-histogram creation.

This PR adds a check in histogram creation that will skip sorting the
buckets if they are already sorted (using the `sort.IsSorted` function).

Ideally in the case where we do need to sort the buckets we could make a
copy of the buckets to guarantee no race conditions.  This PR is more of
a mitigation than a full-on fix for all cases though.