Use semver dependency on Apache Thrift (#83)
Apache Thrift made a breaking api change in the 0.11 series. This package relies on the ~0.10 API. 
Currently downstream packages using dep do not inherit this package's transitive dependency constraint on apache thrift because dep does not consider revision pins when importing constraints from alternate dependency managers. By using a semver range instead of a direct pin.

Another parallel path is to propose amending Dep to consider Glide revision pins as valid transitive constraints. We can do that unilaterally in our fork of dep but it would be much better if we worked with upstream. This pull request gets us out of the immediate pain and unblocks migration to dep within Uber.