Proposal to create a NewRegistry by default
Currently the default behavior on creating a NewReporter is to use the prometheus DefaultReporter, which can result in, for instance difficult to track down panics due to duplicate metric registry when spinning up two different reporters in separate test cases. I would like to propose either creating a NewRegistry by default, or at least documenting the default behavior of this method.